+++ CHANGELOG.md
@@ -25,6 +25,11 @@ Refs. for badges:
 - http://github.com/wayneashleyberry/terminal-dimensions
 - https://github.com/golangci/golangci-lint
 
+## [1.5.10] - 2022-02-22
+### Changed
+- Fixed `New`, was missing setting `defaultIoWriterLevel`, `fields`, `status`.
+- Changed default `io.Writer` level to `None`.
+
 ## [1.5.9] - 2022-02-21
 ### Added
 - For convenience, conforms with `io.Writer` interface. Default level: `error`. `SetIoWriterLevel` changes the default level.